# 校园二手交易APP/openapi.yaml
openapi: 3.0.3
info:
  title: 校园二手交易 APP API
  version: 1.0.0
  description: |
    Android 客户端使用的校园二手交易平台后端接口。所有定义与 `api.md`、`db_schema.md.md`、`spec.md`、
    《后端 API 接口开发规划文档》保持一致。
servers:
  - url: http://localhost:8080/api/v1
    description: 开发环境
  - url: https://api.campus-trade.com/api/v1
    description: 生产示例
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: OK
        errorCode:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time
        data:
          description: 具体数据载荷
          nullable: true
    PaginatedMeta:
      type: object
      properties:
        page:
          type: integer
        size:
          type: integer
        total:
          type: integer
    UserSummary:
      type: object
      properties:
        id: { type: integer }
        username: { type: string }
        role:
          type: string
          enum: [STUDENT, TEACHER, STAFF]
        email: { type: string }
        phone: { type: string, nullable: true }
        avatar: { type: string, nullable: true }
        school: { type: string, nullable: true }
        emailVerified: { type: boolean }
    Product:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: string, nullable: true }
        price: { type: number, format: double }
        category:
          type: string
          enum: [BOOKS, ELECTRONICS, CLOTHING, SPORTS, DAILY, OTHER]
        categoryId:
          type: integer
          nullable: true
          description: 动态分类ID（可选）
        categoryCode:
          type: string
          nullable: true
          description: 动态分类编码（可选）
        categoryName:
          type: string
          nullable: true
          description: 动态分类名称（可选）
        images:
          type: array
          items: { type: string }
        tags:
          type: array
          items: { type: string }
        status:
          type: string
          enum: [DRAFT, ON_SALE, OFF_SALE, SOLD, DELETED]
        auditStatus:
          type: string
          enum: [PENDING, APPROVED, REJECTED]
        viewCount: { type: integer }
        likeCount: { type: integer }
        contactInfo: { type: string, nullable: true }
        location: { type: string, nullable: true }
        remark: { type: string, nullable: true }
        seller:
          $ref: '#/components/schemas/UserSummary'
        createTime: { type: string, format: date-time }
        favorited: { type: boolean }
...
    CartItem:
      type: object
      properties:
        id: { type: integer }
        quantity: { type: integer }
        product: { $ref: '#/components/schemas/Product' }
        createTime: { type: string, format: date-time, nullable: true }
        updateTime: { type: string, format: date-time, nullable: true }
    PaginatedCartItems:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CartItem'
        meta:
          $ref: '#/components/schemas/PaginatedMeta'
...
paths:
  /auth/register:
    post:
      summary: 用户注册
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password, email, role]
              properties:
                username: { type: string, minLength: 3, maxLength: 20 }
                password: { type: string, minLength: 6, maxLength: 20 }
                email: { type: string, format: email }
                phone: { type: string }
                role:
                  type: string
                  enum: [STUDENT, TEACHER, STAFF]
      responses:
        '200':
          description: 注册成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserSummary'
  /auth/verify-email/request:
    post:
      summary: 发送邮箱验证码
      description: 注册后或验证码过期时重新发送 6 位数字验证码
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        '200': { description: 已发送, content: { application/json: { schema: { $ref: '#/components/schemas/ApiResponse' } } } }
  /auth/verify-email:
    post:
      summary: 邮箱验证
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, verificationCode]
              properties:
                email: { type: string, format: email }
                verificationCode: { type: string }
      responses:
        '200': { description: 激活成功, content: { application/json: { schema: { $ref: '#/components/schemas/ApiResponse' } } } }
  /auth/login:
    post:
      summary: 用户登录
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          accessToken: { type: string }
                          expiresIn: { type: integer }
                          user: { $ref: '#/components/schemas/UserSummary' }
  /auth/reset-password/request:
    post:
      summary: 发送密码重置验证码
      description: 仅邮箱已验证的账号可以请求发送验证码
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        '200': { description: 已发送, content: { application/json: { schema: { $ref: '#/components/schemas/ApiResponse' } } } }
...
  /auth/reset-password:
    post:
      summary: 邮箱重置密码
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, verificationCode, newPassword]
              properties:
                email: { type: string, format: email }
                verificationCode: { type: string }
                newPassword: { type: string, minLength: 6 }
      responses:
        '200': { description: 成功, content: { application/json: { schema: { $ref: '#/components/schemas/ApiResponse' } } } }
...
  /cart/items:
    post:
      summary: 添加商品到购物车
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [productId]
              properties:
                productId: { type: integer }
                quantity:
                  type: integer
                  minimum: 1
                  description: 为空时默认 1
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CartItem'
    get:
      summary: 查看购物车列表
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: size
          schema: { type: integer, default: 10 }
      responses:
        '200':
          description: 列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PaginatedCartItems'
    delete:
      summary: 清空购物车
      responses:
        '200': { description: 成功, content: { application/json: { schema: { $ref: '#/components/schemas/ApiResponse' } } } }
  /cart/items/{id}:
    patch:
      summary: 更新购物车商品数量
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [quantity]
              properties:
                quantity:
                  type: integer
                  minimum: 1
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CartItem'
    delete:
      summary: 移除指定购物车条目
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200': { description: 成功, content: { application/json: { schema: { $ref: '#/components/schemas/ApiResponse' } } } }
  /cart/items/product/{productId}:
    delete:
      summary: 根据商品 ID 移除购物车条目
      parameters:
        - in: path
          name: productId
          required: true
          schema: { type: integer }
      responses:
        '200': { description: 成功, content: { application/json: { schema: { $ref: '#/components/schemas/ApiResponse' } } } }
...